+++ libs/reader/include/metaldb/reader/RawTable.hpp
@@ -13,16 +13,16 @@
 namespace metaldb::reader {
     class RawTable {
     public:
-        RawTable(std::vector<char>&& buffer, const std::vector<std::size_t>& rowIndexes, const std::vector<std::string>& columns);
+        RawTable(std::vector<char>&& buffer, const std::vector<std::uint16_t>& rowIndexes, const std::vector<std::string>& columns);
 
         static RawTable invalid();
 
         std::vector<char> data;
         std::vector<std::string> columns;
-        std::vector<std::size_t> rowIndexes;
+        std::vector<std::uint16_t> rowIndexes;
 
         std::string debugStr() const;
-        std::size_t numRows() const;
+        std::uint32_t numRows() const;
         std::string readRow(std::size_t row) const;
 
         bool isValid() const;
